% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendData.R
\name{sendData}
\alias{sendData}
\title{Send Data}
\usage{
sendData(dataframes_list)
}
\arguments{
\item{dataframes_list}{list of data to send}
}
\value{
output
}
\description{
Sends data via email
}
\details{
This function can be used to do some things
}
